+++ khoros/api.py
@@ -6,7 +6,7 @@
 :Example:           ``json_response = khoros.api.get_request_with_retries(url, auth_dict=khoros.auth)``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     02 Oct 2021
+:Modified Date:     23 May 2022
 """
 
 import json
@@ -180,7 +180,6 @@ def _display_ssl_verify_warning():
 
         # Suppress warnings when performing API calls without verifying SSL certificates
         urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-    return
 
 
 def should_verify_tls(khoros_object=None):
@@ -442,7 +441,6 @@ def _report_failed_attempt(_exc_msg, _request_type, _retries):
     _error_msg = f"The {_request_type.upper()} request has failed with the following exception: " + \
                  f"{_exc_name}: {_exc_msg} {_current_attempt}"
     errors.handlers.eprint(f"{_error_msg}\n{_exc_name}: {_exc_msg}\n")
-    return
 
 
 def _raise_exception_for_repeated_timeouts():
@@ -1338,7 +1336,6 @@ def _confirm_field_supplied(_fields_dict):
             break
     if not _field_supplied:
         raise errors.exceptions.MissingRequiredDataError("At least one field must be enabled to retrieve a response.")
-    return
 
 
 def _normalize_base_url(_base_url):